DCS Inverted
Implemented in display strings, channel details, and tone scanning.

Inverted DCS is really just DCS with mark and space in the data stream
inverted. It can be important to have this because older implementations
didn't agree on the meaning of mark and space, so some radios or
repeaters could call normal inverted and vice versa.

Because of how the math works out, inverted codes can be expressed as
non-standard codes. This also means that some non-standard codes are
actually inverted normal codes.