Tolerate some loss of CTCSS / DCS in Rx
Based on #670 by Riku / OH1E and should help the following:
https://www.opengd77.com/viewtopic.php?f=11&t=1179
https://www.opengd77.com/viewtopic.php?f=11&t=794&start=40

I tested using three radios, one with DCS Tx setting, another with no Tx
tone or code, and the GD-77 with this patch receiving. With this patch
the receiving radio will receive if there's a clean DCS signal from the
one radio, and will hold the Rx open just a bit if the signal is doubled
by the second radio or if the first (DCS Tx) radio stops transmitting
but the second is still transmitting. It stops immediately if the signal
is lost.